LCM added: the function should returns the Least Common Multiple of two or more supplied numbers
