+++ math/README.md
@@ -11,5 +11,5 @@ The tables below list all the current built-in Microsoft Excel math functions, g
 | [x] ABS					| Returns the absolute value (ie. the modulus) of a supplied number  |
 | [x] SIGN					| Returns the sign (+1, -1 or 0) of a supplied number			     |
 | [x] GCD					| Returns the Greatest Common Divisor of two or more supplied numbers|
-| [ ] LCM					| Returns the Least Common Multiple of two or more supplied numbers  |
+| [x] LCM					| Returns the Least Common Multiple of two or more supplied numbers  |
